mshv: Add integration test for mshvlog service
Add an functional integration test for mshvlog service. If mshv_diag
module is loaded, mshlog.service should run without any errors.

Move tests in mshv_root_tests.py to appropriately named
mshv_root_stress_tests.py file.
